+++ test/test.js
@@ -3,19 +3,24 @@ var http = require('http')
 var methodOverride = require('..')
 var request = require('supertest')
 
-describe('methodOverride()', function(){
-  var server
-  before(function () {
-    server = createServer()
-  })
-
+describe('methodOverride(getter)', function(){
   it('should not touch the method by default', function(done){
+    var server = createServer()
     request(server)
     .get('/')
     .expect('X-Got-Method', 'GET', done)
   })
 
   describe('with body', function(){
+    it('should work missing body', function(done){
+      var server = createServer('_method')
+
+      request(server)
+      .post('/')
+      .set('Content-Type', 'application/json')
+      .expect('X-Got-Method', 'POST', done)
+    })
+
     it('should be case in-sensitive', function(done){
       var server = createServer('_method', {
         _method: 'DELete'
@@ -81,7 +86,14 @@ describe('methodOverride()', function(){
   describe('with header', function(){
     var server
     before(function () {
-      server = createServer()
+      server = createServer('X-HTTP-Method-Override')
+    })
+
+    it('should work missing header', function(done){
+      request(server)
+      .post('/')
+      .set('Content-Type', 'application/json')
+      .expect('X-Got-Method', 'POST', done)
     })
 
     it('should be case in-sensitive', function(done){
@@ -106,6 +118,38 @@ describe('methodOverride()', function(){
       .set('Content-Type', 'application/json')
       .set('X-HTTP-Method-Override', 'DELETE, PUT')
       .expect('X-Got-Method', 'DELETE', done)
+    })
+  })
+
+  describe('with function', function(){
+    var server
+    before(function () {
+      server = createServer(function(req){
+        return req.headers['x-method-override'] || 'PaTcH'
+      })
+    })
+
+    it('should work missing header', function(done){
+      request(server)
+      .post('/')
+      .set('Content-Type', 'application/json')
+      .expect('X-Got-Method', 'PATCH', done)
+    })
+
+    it('should be case in-sensitive', function(done){
+      request(server)
+      .post('/')
+      .set('Content-Type', 'application/json')
+      .set('X-Method-Override', 'DELete')
+      .expect('X-Got-Method', 'DELETE', done)
+    })
+
+    it('should ignore invalid methods', function(done){
+      request(server)
+      .post('/')
+      .set('Content-Type', 'application/json')
+      .set('X-Method-Override', 'BOGUS')
+      .expect('X-Got-Method', 'POST', done)
     })
   })
 })