+++ History.md
@@ -1,3 +1,10 @@
+unreleased
+==========
+
+  * New interface, less magic
+    - Can specify what header to look for override in, if wanted
+    - Can specify custom function to get method from request
+
 1.0.2 / 2014-05-22
 ==================